Conflicts solved

+++ src/main/java/uz/pdp/simline/dto/respone/PlanDto.java
@@ -12,7 +12,7 @@ public class PlanDto {
     private UUID id;
     private String name;
     private Long expiry;
-    private Double mb;
+    private Long mb;
     private Long sms;
     private Long minute;
     private Double price;
